more rhel fixes to install-dependences

+++ Makefile
@@ -115,8 +115,9 @@ ifeq ($(shell uname), Darwin)
 	port install asciidoc
 	easy_install ply
 else
-	PKGS="libgmp10 strace python-ply python-gevent" if [ -f /usr/bin/yum ] ; then yum install $PKGS; else apt-get install $PKGS; fi
-	ln -sf /usr/lib/x86_64-linux-gnu/libgmp.so /usr/lib/x86_64-linux-gnu/libgmp.so.3
+	if [ -f /usr/bin/yum ] ; then yum install gmp strace python-ply python-gevent; else apt-get install libgmp10 strace python-ply python-gevent; fi
+	if [ -f /usr/lib/x86_64-linux-gnu/libgmp.so ] ; then ln -sf /usr/lib/x86_64-linux-gnu/libgmp.so /usr/lib/x86_64-linux-gnu/libgmp.so.3 ; fi
+	if [ -f /usr/lib64/libgmp.so.10] ; then ln -s /usr/lib64/libgmp.so.10 /usr/lib64/libgmp.so.3; fi
 endif
 
 install-python-example-dependences:
